A rendering pipeline implementation for the GraphicsContext is in progress.

+++ Sources/DKGame/Math/Scalar.swift
@@ -2,13 +2,13 @@
 //  File: Scalar.swift
 //  Author: Hongtae Kim (tiff2766@gmail.com)
 //
-//  Copyright (c) 2022 Hongtae Kim. All rights reserved.
+//  Copyright (c) 2022-2023 Hongtae Kim. All rights reserved.
 //
 
 import Foundation
 
-public typealias Scalar = Float32
-//public typealias Scalar = CGFloat
+//public typealias Scalar = Float32
+public typealias Scalar = Float64
 
 
 extension Comparable {
